stage3

+++ laravel/app/Http/Controllers/Admin/UsersController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\User;
 use App\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UsersController extends Controller
 {
@@ -66,6 +67,11 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        if(Gate::denies('edit-users'))
+        {
+            return redirect()->route('admin.users.index');
+        }
+
         $roles = Role::all();
 
         return view('admin.users.edit', [
@@ -84,6 +90,9 @@ class UsersController extends Controller
     public function update(Request $request, User $user)
     {
         $user->roles()->sync($request->roles);
+        $user->name = $request->name;
+        $user->email = $request->email;
+        $user->save();
 
         return redirect()->route('admin.users.index');
     }
@@ -96,6 +105,10 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if(Gate::denies('delete-users'))
+        {
+            return redirect()->route('admin.users.index');
+        }
         $user->roles()->detach();
         $user->delete();
